+++ Source/BitBucket.REST.API/Models/Standard/Inline.cs
@@ -5,10 +5,10 @@ namespace BitBucket.REST.API.Models.Standard
     public class Inline
     {
         [JsonProperty(PropertyName = "to")]
-        public int? To { get; set; }
+        public long? To { get; set; }
 
         [JsonProperty(PropertyName = "from")]
-        public int? From { get; set; }
+        public long? From { get; set; }
 
         [JsonProperty(PropertyName = "path")]
         public string Path { get; set; }
